+++ tests/Neo.Plugins.L2Batch.UnitTests/Usings.cs
@@ -2,5 +2,6 @@ global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Neo;
 global using Neo.L2;
 global using Neo.L2.Batch;
+global using Neo.L2.Bridge;
 global using Neo.L2.Telemetry;
 global using Neo.Plugins.L2;